Move shape_mesh.h from components/ to systems/ (refs #1194)
ShapeProps and ShapeMeshConfig are render-resource configuration, not
entity component data — they're emplaced into registry.ctx() and consumed
only by camera_system. Per the #1194 audit verdict (MOVE), relocate the
header beside its owning system.

Sites updated:
- app/systems/camera_system.cpp: relative include
- tests/test_perspective.cpp: "components/" -> "systems/"
- tests/test_audio_offset_calibration.cpp: "components/" -> "systems/"

Validation: 901 Catch2 tests / 2957 assertions pass; zero warnings on
both unity and non-unity Clang -Wall -Wextra -Werror builds.
